Move shapes to separate module
